+++ src/app/admin/admin.module.ts
@@ -1,13 +1,14 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-
+import { SharedModule } from '../shared/shared.module';
+import { AdminServicesModule } from './admin-services.module';
 import { AdminRoutingModule } from './admin-routing.module';
 
 @NgModule({
   declarations: [AdminRoutingModule.components
   ],
   imports: [
-    CommonModule,
+    SharedModule,
+    AdminServicesModule,
     AdminRoutingModule
   ]
 })