+++ src/app/admin/components/index.ts
@@ -1,3 +1,5 @@
 export * from './admin-dashboard/admin-dashboard.component';
 export * from './manage-products/manage-products.component';
+export * from './manage-product/manage-product.component';
+export * from './manage-product-form/manage-product-form.component';