+++ src/app/shared/shared.module.ts
@@ -4,10 +4,12 @@ import { FormsModule } from '@angular/forms';
 import { HighlightItemDirective } from './directives/highlight-item.directive';
 import { ClickItemDirective } from './directives/click-item.directive';
 import { OrderByPipe } from './pipes/order-by.pipe';
+import { ProductComponent, ProductFormComponent} from './../products';
 
 
 const dirs = [HighlightItemDirective, ClickItemDirective, OrderByPipe];
 const mods = [CommonModule, FormsModule];
+
 
 @NgModule({
   imports: [...mods],